Make package structure match assertj conventions
