+++ src/main/java/module-info.java
@@ -17,8 +17,7 @@
  * Main AssertJ Eclipse Collections Module
  */
 module org.assertj.eclipse.collections {
-  exports org.assertj.eclipse.collections;
-  exports org.assertj.eclipse.collections.multimap;
+  exports org.assertj.eclipse.collections.api;
   exports org.assertj.eclipse.collections.error;
 
   requires org.assertj.core;
