+++ src/main/java/org/assertj/eclipse/collections/api/SoftAssertions.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.assertj.eclipse.collections;
+package org.assertj.eclipse.collections.api;
 
 import org.assertj.core.api.AbstractSoftAssertions;
 import org.assertj.core.api.SoftAssertionsProvider;
